+++ crates/rollup-interface/src/state_machine/mod.rs
@@ -105,6 +105,7 @@ pub trait BasicAddress:
     + core::fmt::Debug
     + core::fmt::Display
     + Send
+    + Unpin
     + Sync
     + Clone
     + core::hash::Hash
@@ -153,6 +154,8 @@ pub struct StateUpdateInfo<StfState> {
     pub ledger_reader: rockbound::cache::delta_reader::DeltaReader,
     /// What the next event number will be after the state update.
     pub next_event_number: u64,
+    /// What the next transaction number will be after the state update.
+    pub next_tx_number: u64,
     /// The slot number of the rollup following the state update.
     pub slot_number: SlotNumber,
     /// The latest slot number that was finalized.